# Rust build image optimizations