--- a/packages/cli/src/lib/defaults/build-images/wasm/rust/Dockerfile.mustache
+++ b/packages/cli/src/lib/defaults/build-images/wasm/rust/Dockerfile.mustache
@@ -32,6 +32,8 @@ WORKDIR /linked-packages
 COPY {{dir}} ./{{name}}
 {{/polywrap_linked_packages}}
 
+WORKDIR /
+RUN USER=root cargo new --lib project
 WORKDIR /project
 
 # Copy all manifest files
@@ -39,13 +41,8 @@ WORKDIR /project
 COPY {{.}} .
 {{/polywrap_manifests}}
 
-# Copy all source files
-{{#include}}
-COPY {{.}} {{.}}
-{{/include}}
-{{#polywrap_module}}
-COPY {{dir}} {{dir}}
-{{/polywrap_module}}
+# Copy Cargo manifest files
+COPY ./{{dir}}/Cargo.toml ./{{dir}}/Cargo.toml
 
 {{#polywrap_linked_packages.length}}
 # Link any local packages
@@ -58,15 +55,10 @@ RUN PACKAGE_NAME={{name}}; \
     {{/polywrap_module}}
     true
 {{/polywrap_linked_packages}}
-
 {{/polywrap_linked_packages.length}}
 
 {{#polywrap_module}}
-# Remove any Cargo.lock files
-RUN rm -rf {{dir}}/Cargo.lock
-
-# Ensure the Wasm module is configured to use imported memory
-ENV RUSTFLAGS="-C link-arg=-z -C link-arg=stack-size=65536 -C link-arg=--import-memory"
+# Mock build for dependencies:
 
 # Ensure the module at {{dir}} has the crate-type = ["cdylib"]
 RUN toml set ./{{dir}}/Cargo.toml lib.crate-type ["cdylib"] > ./{{dir}}/Cargo-local.toml && \
@@ -83,13 +75,58 @@ RUN toml set ./{{dir}}/Cargo.toml package.name "module" > ./{{dir}}/Cargo-local.
     mv ./{{dir}}/Cargo-local.toml ./{{dir}}/Cargo.toml && \
     true
 
-# Make the build directory
-RUN rm -rf ./build
-RUN mkdir ./build
+RUN cargo build --manifest-path ./{{dir}}/Cargo.toml --release
+
+# Copy all source files
+{{#include}}
+COPY {{.}} {{.}}
+{{/include}}
+COPY {{dir}} {{dir}}
+RUN rm ./target/release/deps/module*
+
+{{/polywrap_module}}
+
+{{#polywrap_linked_packages.length}}
+# Link any local packages
+{{#polywrap_linked_packages}}
+RUN PACKAGE_NAME={{name}}; \
+    {{#polywrap_module}}
+    toml set ./{{dir}}/Cargo.toml dependencies.$PACKAGE_NAME.path /linked-packages/$PACKAGE_NAME > ./{{dir}}/Cargo-local.toml && \
+    rm -rf ./{{dir}}/Cargo.toml && \
+    mv ./{{dir}}/Cargo-local.toml ./{{dir}}/Cargo.toml && \
+    {{/polywrap_module}}
+    true
+{{/polywrap_linked_packages}}
+{{/polywrap_linked_packages.length}}
+
+{{#polywrap_module}}
+# Actual build:
+
+# Ensure the module at {{dir}} has the crate-type = ["cdylib"]
+RUN toml set ./{{dir}}/Cargo.toml lib.crate-type ["cdylib"] > ./{{dir}}/Cargo-local.toml && \
+    rm -rf ./{{dir}}/Cargo.toml && \
+    mv ./{{dir}}/Cargo-local.toml ./{{dir}}/Cargo.toml && \
+    true
+
+# Clean up artifacts left by the toml CLI program ("["cdylib"]" -> ["cdylib"])
+RUN sed -i 's/"\[cdylib\]"/\["cdylib"\]/g' ./{{dir}}/Cargo.toml
+
+# Ensure the package name = "module"
+RUN toml set ./{{dir}}/Cargo.toml package.name "module" > ./{{dir}}/Cargo-local.toml && \
+    rm -rf ./{{dir}}/Cargo.toml && \
+    mv ./{{dir}}/Cargo-local.toml ./{{dir}}/Cargo.toml && \
+    true
+
+# Ensure the Wasm module is configured to use imported memory
+ENV RUSTFLAGS="-C link-arg=-z -C link-arg=stack-size=65536 -C link-arg=--import-memory"
 
 # Build the module at {{dir}}
 RUN cargo build --manifest-path ./{{dir}}/Cargo.toml \
     --target wasm32-unknown-unknown --release
+
+# Make the build directory
+RUN rm -rf ./build
+RUN mkdir ./build
 
 # Enable the "WASM_INTERFACE_TYPES" feature, which will remove the __wbindgen_throw import.
 # See: https://github.com/rustwasm/wasm-bindgen/blob/7f4663b70bd492278bf0e7bba4eeddb3d840c868/crates/cli-support/src/lib.rs#L397-L403