--- a/packages/js/client/src/__tests__/e2e/wasm-rs.spec.ts
+++ b/packages/js/client/src/__tests__/e2e/wasm-rs.spec.ts
@@ -1,6 +1,7 @@
 import {
   PolywrapClientConfig,
   createPolywrapClient,
+  PolywrapClient,
 } from "../..";
 import * as TestCases from "./test-cases";
 import { makeMemoryStoragePlugin } from "./memory-storage";
@@ -12,6 +13,10 @@ import {
   providers
 } from "@polywrap/test-env-js";
 import { GetPathToTestWrappers } from "@polywrap/test-cases";
+import fse from "fs-extra";
+import path from "path";
+import { execSync } from "child_process";
+const { performance } = require("perf_hooks");
 
 jest.setTimeout(1200000);
 
@@ -330,3 +335,101 @@ describe("wasm-rs test cases", () => {
     );
   })
 });
+
+describe.only("Wasm-rs benchmarking", () => {
+  const wrapperPath = `${GetPathToTestWrappers()}/wasm-rs/bignumber-type`
+  const wrapperUri = `fs/${wrapperPath}/build`
+
+  let cacheFiles = new Map<string, string>();
+  const mockFunc = `
+  fn froo() -> &'static str {
+"foo"
+}
+  `;
+
+  const modifySource = () => {
+    const libPath = path.join(wrapperPath, "src", "lib.rs");
+    const libFile = fse.readFileSync(libPath, "utf-8");
+
+    cacheFiles.set(libPath, libFile);
+
+    const modifiedFile = `${libFile}\n${mockFunc}`;
+
+    fse.writeFileSync(libPath, modifiedFile);
+  };
+
+  const buildImage = async (name: "current" | "new"): Promise<number> => {
+    const startTime = performance.now();
+
+    await buildWrapper(
+      wrapperPath,
+      name === "current"? "./polywrap-current.yaml": "./polywrap.yaml"
+    );
+
+    const endTime = performance.now();
+    const msTime = endTime - startTime;
+
+    //Make sure the wrapper works correctly
+    await TestCases.runBigNumberTypeTest(
+      new PolywrapClient(), wrapperUri
+    );
+
+    return msTime
+  };
+
+  beforeEach(() => {
+    fse.removeSync(`${wrapperPath}/build`);
+    fse.removeSync(`${wrapperPath}/.polywrap`);
+  });
+
+  const restoreSource = () => {
+    for (const [key, value] of cacheFiles) {
+      fse.writeFileSync(key, value);
+    }
+  };
+
+  it("Build image performance", async () => {
+    //Delete cached images and containers
+    execSync(`docker system prune -a -f`);
+    
+    //Build the wrapper with no previously cached images
+    const firstBuildTimeNew = await buildImage("new");
+    console.log(`1st build - no cache (new): ${firstBuildTimeNew.toFixed(2)}ms`);
+
+    //Build the wrapper again
+    const secondBuildTimeNew = await buildImage("new");
+    console.log(`2nd build - with cache (new): ${secondBuildTimeNew.toFixed(2)}ms`);
+
+    //Modify the source code and measure build time
+    modifySource();
+
+    const timeAfterSourceNew = await buildImage("new",);
+    console.log(`3rd build - modified source (new): ${timeAfterSourceNew.toFixed(2)}ms`);
+
+    restoreSource();
+
+    // Repeat the process for current image and compare
+    execSync(`docker system prune -a -f`);
+    
+    const firstBuildTimeCurrent = await buildImage("current");
+    console.log(`1st build - no cache (current): ${firstBuildTimeCurrent.toFixed(2)}ms`);
+
+    const secondBuildTimeCurrent = await buildImage("current");
+    console.log(`2nd build - with cache (current): ${secondBuildTimeCurrent.toFixed(2)}ms`);
+
+    modifySource();
+
+    const timeAfterSourceCurrent = await buildImage("current");
+    console.log(`3rd build - modified source (current): ${timeAfterSourceCurrent.toFixed(2)}ms`);
+
+    restoreSource();
+
+    expect(timeAfterSourceNew).toBeGreaterThan(timeAfterSourceCurrent);
+
+    await new Promise<void>((res) => {
+      setTimeout(() => {
+        res();
+      }, 100000);
+    })
+  })
+})