--- a/packages/test-cases/cases/wrappers/wasm-rs/benchmarks/Cargo.toml
+++ b/packages/test-cases/cases/wrappers/wasm-rs/benchmarks/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "module"
+version = "0.1.0"
+description = "Query module of bigint-type e2e test"
+authors = [
+  "Kobby Pentangeli <kobbypentangeli@gmail.com>",
+  "Jordan Ellis <jelli@dorg.tech>"
+]
+repository = "https://github.com/polywrap/monorepo"
+license = "MIT"
+edition = "2021"
+
+[dependencies]
+polywrap-wasm-rs = { path = "../../../../../../wasm/rs" }
+serde = { version = "1.0", features = ["derive"] }
+ethers = "0.14.0"
+
+[lib]
+crate-type = ["cdylib"]
+
+[profile.release]
+opt-level = 's'
+lto = true
+panic = 'abort'