--- a/packages/test-cases/cases/wrappers/wasm-rs/benchmarks/polywrap-current.build.yaml
+++ b/packages/test-cases/cases/wrappers/wasm-rs/benchmarks/polywrap-current.build.yaml
@@ -1,0 +1,7 @@
+format: 0.1.0
+docker:
+  name: benchmark-type-wasm-rs
+  dockerfile: ./Dockerfile
+linked_packages:
+  - name: polywrap-wasm-rs
+    path: ../../../../../wasm/rs