--- a/packages/test-cases/cases/wrappers/wasm-rs/benchmarks/polywrap.build.yaml
+++ b/packages/test-cases/cases/wrappers/wasm-rs/benchmarks/polywrap.build.yaml
@@ -1,0 +1,6 @@
+format: 0.1.0
+docker:
+  name: benchmark-type-wasm-rs
+linked_packages:
+  - name: polywrap-wasm-rs
+    path: ../../../../../wasm/rs