--- a/packages/test-cases/cases/wrappers/wasm-rs/benchmarks/polywrap.yaml
+++ b/packages/test-cases/cases/wrappers/wasm-rs/benchmarks/polywrap.yaml
@@ -1,0 +1,6 @@
+format: 0.1.0
+name: Benchmark
+language: wasm/rust
+build: ./polywrap.build.yaml
+schema: ./schema.graphql
+module: ./Cargo.toml