--- a/packages/test-cases/cases/wrappers/wasm-rs/benchmarks/schema.graphql
+++ b/packages/test-cases/cases/wrappers/wasm-rs/benchmarks/schema.graphql
@@ -1,0 +1,12 @@
+type Module {
+  method(
+    arg1: BigNumber!
+    arg2: BigNumber
+    obj: BigNumberArg!
+  ): BigNumber!
+}
+
+type BigNumberArg {
+  prop1: BigNumber!
+  prop2: BigNumber
+}